Added a first stab at the CrowdLogger Remote Module framework.
"Sandboxing" is done using iframes and postMessage. There are two API layers: the CrowdLogger-side (CLI) and the CrowdLogger Remote Module-side (CLRMI). CLRMs interact with the CLRMI, which then talks to the CLI, which has access to the rest of the CrowdLogger codebase. CLRMs can be imported from URLs or from strings (e.g., that are stored in a database).